build: create domain and view package and relocate files into them
